airflow/requirements.txt
apache-airflow-providers-snowflake
astronomer-cosmos
dbt-snowflake
boto3
pandas
snowflake-connector-python